pyproject.toml
[project]
name = "cropfresh-service-ai"
version = "0.1.0"
description = "CropFresh AI Service - Agentic RAG System for Agricultural Marketplace"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [
    { name = "CropFresh Team" }
]

dependencies = [
    # ═══════════════════════════════════════════════════════════════
    # LLM & Agent Framework
    # ═══════════════════════════════════════════════════════════════
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-community>=0.3.0",
    "groq>=0.11.0",
    # ═══════════════════════════════════════════════════════════════
    # Vector Database & Embeddings
    # ═══════════════════════════════════════════════════════════════
    "qdrant-client>=1.12.0",
    "sentence-transformers>=3.0.0",
    # ═══════════════════════════════════════════════════════════════
    # API Framework
    # ═══════════════════════════════════════════════════════════════
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "grpcio>=1.68.0",
    "grpcio-tools>=1.68.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "python-multipart>=0.0.9",
    # ═══════════════════════════════════════════════════════════════
    # HTTP & Async
    # ═══════════════════════════════════════════════════════════════
    "httpx>=0.28.0",
    "aiohttp>=3.11.0",
    "websockets>=12.0",
    # ═══════════════════════════════════════════════════════════════
    # Data Processing
    # ═══════════════════════════════════════════════════════════════
    "numpy>=2.0.0",
    "pandas>=2.2.0",
    "pillow>=10.4.0,<11.0.0",
    # ═══════════════════════════════════════════════════════════════
    # Caching & Storage
    # ═══════════════════════════════════════════════════════════════
    "redis>=5.2.0",
    "neo4j>=5.15.0",
    # ═══════════════════════════════════════════════════════════════
    # Utilities
    # ═══════════════════════════════════════════════════════════════
    "python-dotenv>=1.0.0",
    "loguru>=0.7.0",
    "tenacity>=9.0.0",
    "orjson>=3.10.0",
    "postgrest>=2.27.1",
    "neo4j>=6.0.3",
    "certifi>=2026.1.4",
    "pip-system-certs>=5.3",
    "langsmith>=0.6.2",
    "cohere>=5.20.1",
    "lxml>=5.3.0,<6.0.0",              # Compatible with crawl4ai
]

[project.optional-dependencies]
# Machine Learning / Deep Learning
ml = [
    "torch>=2.5.0",
    "torchvision>=0.20.0",
    "torchaudio>=2.5.0",
    "transformers>=4.46.0",
    "accelerate>=1.2.0",
]

# Vision Models
vision = [
    "ultralytics>=8.3.0",          # YOLOv11/12
    "opencv-python>=4.10.0",
    "supervision>=0.25.0",
]

# Voice / Audio Processing
voice = [
    "soundfile>=0.12.0",
    "edge-tts>=6.1.0",
    "gtts>=2.5.0",
    "pydub>=0.25.0",
    "onnx>=1.16.0",
    "onnxruntime>=1.18.0",
]

# Advanced Voice (WebRTC, VAD, Streaming)
voice-advanced = [
    "aiortc>=1.9.0",            # WebRTC for Python
    "av>=12.0.0",               # Audio/Video processing (for aiortc)
    "aioice>=0.9.0",            # ICE for WebRTC
]
# Note: transformers, torch are in [ml] extras - used for Whisper + IndicTTS

# Web Scraping & Browsing (Phase 10)
web = [
    "playwright>=1.49.0",
    "crawl4ai>=0.4.0",
    "fake-useragent>=1.5.0",
    "trafilatura>=1.14.0",
    "beautifulsoup4>=4.12.0",
    "feedparser>=6.0.0",        # RSS feed parsing
    "mcp>=1.0.0",                # Model Context Protocol SDK
]

# Observability
observability = [
    "opentelemetry-api>=1.28.0",
    "opentelemetry-sdk>=1.28.0",
    "opentelemetry-exporter-otlp>=1.28.0",
    "prometheus-client>=0.21.0",
]

# Development Tools
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "ipython>=8.30.0",
    "jupyter>=1.1.0",
]

# All dependencies
all = [
    "cropfresh-service-ai[ml,vision,voice,web,observability,dev]"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --tb=short"

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
